requirements.txt
Django==4.2.1
psycopg2_binary==2.9.6
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
pillow==9.5.0
django-cors-headers==4.0.0
django-redis==5.2.0
